Fixed ability change bug
No longer set power->powerCooldown.second when an ability changes; this should be done at the end of the method so that the current charges are properly capped if needed.

When switching to an ability that has limited charges, if the current ability does not have a charge limit then the current charges will be reset to initialCharges. This behaviour is temporary and will be modified once ability groups are added.