Wrapper class around xml_document
Added new class that inherits rapidxml::xml_document but also manages the string containing the actual xml.

It also includes a parse_file method that will automatically assign the contents of the named resource file to the document.

The rapidxml::xml_document class itself is not directly accessible (except as a member of an xml node or attribute), and if accessed its methods cannot be called.