Properly excluded Discord from the build if disabled (#276)
Previously the Discord library was only searched and linked
on Windows builds. But due to the include list of DiscordIntegration.h
a linux build required the discord-rpc header file to be present for a
successful compilation.

Now the discord library can be enabled or disabled via a CMake option.
According to the previous behaviour this option is only enabled by
default on Windows builds.